Consolidate plotting functions into a single module
